config.toml: hardcode baseURL and canonifyURL options
Ensures that the site when hosted on Github Pages
properly loads assets and images generated after invoking:

    hugo

On other platforms though, without these configurations,
they'll load alright, so perhaps we need to add a mental
TODO to investigate why we need these options.

Thanks to @smg727 for helping with figuring this out!